Ok, the WalkerTree concept is good save for two things:
1) The algorithm to create the right tree is very difficult.
   Currently it still needs to collapse parent nodes with the same
   start point, constructing a tree that would be more shallow.
2) It's computationally expensive. On the files I'm building it on,
   it doesn't seem to be a problem, although the computer battery
   is running suspiciously low, and the OS reports iterm is using a
   lot of power.

Even though I just started with the walker nodes, I'm wondering, maybe
a better approach would be to just go for what I'm looking for -
actually look at the code rather than the AST. After all, the ast is
actually an intermediate representation, a proxy, for the code I'm
trying to get at. And it'd be computationally simple.