Highlight farthest coinciding ancestor
Up until now, the node while going Up and Down would not be highlighted
consistently with those you travel to while browsing Left and Right.

In simple words, you may only be "selecting" (understand: highlighting)
a keyword like `if` when you'd expect to "select" the entire if-block.

This commit remedies that issue by selecting the farthest ancestor of
the candidate node that shares its starting coordinates.