Avoid highlighting 'block'-like nodes
When jumping to the first statement of a code block, in some languages,
the attempt at highlighting the outermost ancestor that shares the same
starting coordinates would erroneously highlight the entire code block.

The same may happen for the first statement of a file, which would
previously highlight the entire root node, and consequently: the entire
file.