Iterate on highlighting scheme
Get some polish on the highlighting -- abstract where needed, blend,
    meld, mix, and incorporate other parts of the codebase.

Major change here is using regex matchers on highlight target black
list.

Works on my machine, that is to say, on the languages I have here so
far.

Next step is to test more filetypes - got errors about not having the
right TS parsers installed for the languages in the fixtures. So
either gotta bootstrap more parsers into the minimal test init, or
just go with what's built in to nvim, which I think includes C.