Treewalker movements now add to the jumplist
Similar to { and }, Treewalker movements now register in the built
in jumplist (:help jumplist).

The goal is so large erroneous movements are easier to recover from.
Also, using it for a bit, it just feels more natural. Since { and }
do it too, and those are pretty similar in scope to Treewalker, I figure
it makes sense to include.

Closes #7