Notify and refuse to run when there's a missing parser
When the current instance is missing the current language's LSP parser,
this plugin will not work. Currently a big ugly error message happens.
With this change, instead, a one time notification will be made telling
the user Treewalker won't work without it, and calling the plugin will
have no effect until a parser is installed.

Closes #12

Co-authored-by: BBBoy01 <1156678002@qq.com>