Change default highlight_group
As I changed colorschemes, I started to notice that CursorLine was the
default color I was originally going for. ColorColumn happened to be the
right one on the colorschemes I had been using, but it wasn't right
across a broader spectrum of colorschemes.

This change may introduce a change to users; I hope it isn't too
jarring. My rationale is twofold:

* the readme suggests using the plugin with hard coded defaults, so many
  users may be doing this, which means this change won't affect them
* CursorLine is the same color as ColorColumn for many colorschemes

Hopefully this ends up being an improvement.