Implement lateral swapping
There has been a good amount of noise about wanting lateral
swapping. This would achieve parody with some of the older plugins,
syntax-tree-surfer and tree-climber. To have this be built in and
just work in the Treewalker way seems valuable.

With some borrowed (but heavily modified) code from
nvim-treesitter.ts_utils, this was pretty easy to achieve.

My only concern is is that right/left swapping, on top level nodes
like definitions and declarations, does kind of the same thing as
up/right swapping, but worse. Hopefully that doesn't pollute the
feel of the plugin.

Addresses #16 and #4
Duplicates #10