Swap nodes even when cursor is inside string
But actually, more generally, this uses a strategy to find current/next
nodes during lateral swapping. Namely, if it can't find next swap
targets on a current node, it will look for one from the parent, and then treat
the current node as parent if it has a sibling. This is repeated until
something to swap is found.

Closes #18