Add unsolved rust test case
In Rust, there's a certain pattern that doesn't get swapped right (see
the spec). I scratched my head for a while trying to solve it, then
noticed that :TSTexObjectSwap was having the same problem! So I'm
satisfied adding it as a myster.

TL;DR lateral swapping is finicky.

It works _pretty good_, but not _really good_. But it doesn't look like
any of them do.