Enhancement: allow for up/down swapping from beginning of indented lines
At the time of writing, it seems possible (although not certain)
that #22 is from the lack of this feature.

Now up/down swaps will work if you're on a line that can be up/down
swapped, even if your cursor is at a column number less than where the
code on the line starts.