Enhacement: movement to neighbor nodes don't add to jumplist
The jumplist was getting overpopulated by this plugin. Now jumping to
neighbor nodes doesn't lead to adding to the jumplist. This'll reduce
pollution from treewalker.

The implementation is not perfect - it's a bit tricky to find the
correct nodes to mark when the plugin is stateless, which it will
certainly remain.

This leads to one tradeoff:

When beginning some treewalker movement down a block of one line nodes,
the originating node won't be marked.

But In/Out now add to both the originating node and the destination
node. This helps make up for the aforementioned tradeoff -- if you go
into a function, then down a bunch on all neighboring nodes, then
ctrl-o, you'll go back to the first node inside the function. However if
you just start there without having jumped in with treewalker, that node
won't be registered.

Fair tradeoff to make I think.